fix invalid cursor states and visual selection issues
After deleting lines, the cursor is moved to a valid position consistent with the hjkl
movement limits. E.g., after deleting the last line with dd, it moves up one line. This
furthermore prevents the internal start index of the visual selection to be decreased past
zero and makes delete and yank of the visual selection work when the selection is expanded
from the bottom, i.e., the start is below the end.